Centralise initial layout loading
